add generate_episode to Model and rename file name
