+++ models/base_model.py
@@ -1,8 +1,12 @@
+import random
+
+
 class Model:
-    def __init__(self, state_space, action_space, gamma=0.99):
+    def __init__(self, state_space, action_space, gamma=0.99, epsilon=0.01):
         self.state_space = state_space
         self.action_space = action_space
         self.gamma = gamma
+        self.epsilon = epsilon
         self.policy = self.initialize_policy()
 
     def initialize_policy(self):
@@ -16,3 +20,9 @@ class Model:
 
     def train(self, episodes):
         raise NotImplementedError("This method should be overridden by subclasses")
+
+    def generate_episode(self, env, min_length=4):
+        # Randomly generate the initial state (for each episode) beginning with the first period to period k−4.
+        start = random.randint(0, len(env.data) - min_length)
+        end = random.randint(start + min_length, len(env.data))
+        return [i for i in range(start, end)]
